Outputs resulting properly

+++ main.py
@@ -4,8 +4,8 @@ def main():
 
     total = num_male + num_fem
 
-    m_perc = (num_male / (num_fem + num_male)) *100 
-    f_perc = (num_fem / (num_fem + num_male)) *100
+    m_perc = float((num_male / (num_fem + num_male)) *100)
+    f_perc = float((num_fem / (num_fem + num_male)) *100)
 
     return total, num_male, num_fem, m_perc, f_perc
